feat cascade delete for users

+++ api/lib/src/route_helpers/users/delete_user.dart
@@ -6,6 +6,19 @@ import 'package:tools/tools.dart';
 Future<Response> deleteUser(RequestContext context, Id userId) async {
   await Ensure.isAdmin(context);
 
+  // Get info of the user to be deleted.
+  final getDeletedUserRes = await context.db.collection('users').findOne({'_id': userId});
+  if (getDeletedUserRes == null) {
+    throw NotFound('user');
+  }
+
+  final deletedUser = UserInfo.fromJson(getDeletedUserRes);
+
+  // Ensure that admins are not deleting other admins.
+  if (deletedUser.role.isAdmin && deletedUser.id != (await context.user).id) {
+    throw Forbidden(Errors.permissionDenied);
+  }
+
   // Delete user from the database.
   final deleteUserRes = await context.db.collection('users').deleteOne({'_id': userId});
 
@@ -15,7 +28,26 @@ Future<Response> deleteUser(RequestContext context, Id userId) async {
     throw Exception('Nothing was deleted.');
   }
 
-  // TODO: cascade delete ??
+  // If the deleted user is a student
+  if (deletedUser.role.isStudent) {
+    // Delete this student from every course
+    final deleteStudentRes = await context.db.collection('courses').updateMany(
+      {'studentIds': deletedUser.id},
+      {
+        '\$pull': {'studentIds': deletedUser.id}
+      },
+    );
+    Ensure.hasNoWriteErrors(deleteStudentRes);
+  }
+
+  // If the deleted user is a teacher
+  if (deletedUser.role.isTeacher) {
+    // Delete every course of this teacher.
+    final deleteTeacherCourseRes = await context.db.collection('courses').deleteMany(
+      {'teacherId': deletedUser.id},
+    );
+    Ensure.hasNoWriteErrors(deleteTeacherCourseRes);
+  }
 
   // Send response.
   return Response.json(body: {
